+++ src/base/vgcanvas.c
@@ -25,6 +25,14 @@
 #include "tkc/mem.h"
 #include "tkc/utils.h"
 
+ret_t vgcanvas_set_assets_manager(vgcanvas_t* vg, assets_manager_t* assets_manager) {
+  return_value_if_fail(vg != NULL, RET_BAD_PARAMS);
+
+  vg->assets_manager = assets_manager;
+
+  return RET_OK;
+}
+
 ret_t vgcanvas_reset(vgcanvas_t* vg) {
   return_value_if_fail(vg != NULL && vg->vt->reset != NULL, RET_BAD_PARAMS);